refactor bff fields

+++ src/main/java/az/kapitalbank/bffumico/dto/OrderProductItemDto.java
@@ -1,5 +1,7 @@
 package az.kapitalbank.bffumico.dto;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.AccessLevel;
@@ -10,7 +12,7 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class OrderProductItemDto {
-    Integer productAmount;
+    BigDecimal productAmount;
     String productId;
     String productName;
     String orderNo;
